pyproject.toml
[project]
name = "factchecker-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.43.0",
    "sqlalchemy>=2.0.44",
]
